requirements.txt
groq
numexpr
langchain
langchain_groq
langchain_core
langchain_community
python-dotenv
nbconvert
chainlit